pyproject.toml
[project]
name = "bcn-art-compass-telegram-bot"
version = "0.1.0"
description = "Telegram bot frontend for BCN Art Compass orchestrator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true


